Move more shared headers into inc
