+++ fontforge/scripting.c
@@ -54,8 +54,6 @@
 #include "fvmetrics.h"
 #include "gfile.h"
 #include "gutils.h"
-#include "gutils/prefs.h"
-#include "gutils/unicodelibinfo.h"
 #include "lookups.h"
 #include "macbinary.h"
 #include "mem.h"
@@ -65,6 +63,7 @@
 #include "parsepdf.h"
 #include "parsepfa.h"
 #include "parsettf.h"
+#include "prefs.h"
 #include "print.h"
 #include "savefont.h"
 #include "scriptfuncs.h"
@@ -83,6 +82,7 @@
 #include "tottfgpos.h"
 #include "ttf.h"
 #include "ttfinstrs.h"
+#include "unicodelibinfo.h"
 #include "ustring.h"
 #include "utype.h"
