+++ fontforge/start.c
@@ -33,9 +33,9 @@
 #include "ffglib.h"
 #include "fontforgevw.h"
 #include "gfile.h"
-#include "gutils/unicodelibinfo.h"
 #include "namelist.h"
 #include "psfont.h"
+#include "unicodelibinfo.h"
 
 #include <locale.h>
 #include <sys/time.h>
