+++ fontforgeexe/charview.c
@@ -41,15 +41,14 @@
 #include "fontforgeui.h"
 #include "fvcomposite.h"
 #include "fvfonts.h"
-#include "gdraw/hotkeys.h"
 #include "gkeysym.h"
 #include "gresedit.h"
 #include "gresource.h"
-#include "gutils/prefs.h"
-#include "gutils/unicodelibinfo.h"
+#include "hotkeys.h"
 #include "lookups.h"
 #include "mm.h"
 #include "namelist.h"
+#include "prefs.h"
 #include "sfd.h"
 #include "spiro.h"
 #include "splinefill.h"
@@ -58,6 +57,7 @@
 #include "splinesaveafm.h"
 #include "splineutil.h"
 #include "splineutil2.h"
+#include "unicodelibinfo.h"
 #include "ustring.h"
 #include "utype.h"
 #include "wordlistparser.h"
