+++ fontforgeexe/prefs.c
@@ -38,10 +38,10 @@
 #include "gresedit.h"
 #include "gresource.h"
 #include "groups.h"
-#include "gutils/prefs.h"
 #include "macenc.h"
 #include "namelist.h"
 #include "othersubrs.h"
+#include "prefs.h"
 #include "sfd.h"
 #include "splineutil.h"
 #include "ttf.h"
