+++ inc/unicodelibinfo.h
